* Change to config class * updated unit tests
