+++ test/config/bogus_feed_database_dir.conf
@@ -1,0 +1,22 @@
+[general]
+worker_type=local
+
+; ONLY for worker_type of local
+cb_server_url=https://127.0.0.1:443
+cb_server_token=wFC4he6OBd20IBbB0XkhE5ZsjBQUj1GB
+
+yara_rules_dir=./rules
+
+postgres_host=localhost
+postgres_username=cb
+postgres_password=6PGcbuwlQnIuPqOF
+postgres_db=cb
+postgres_port=5002
+
+niceness=1
+concurrent_hashes=8
+disable_rescan=False
+num_days_binaries=365
+
+; file, not dir
+feed_database_dir={HERE}/README.md
